Rename BuiltinClient -> BuiltinEvaluator
