+++ app/Commands/Dev/Nockma/Run/Options.hs
@@ -1,18 +1,18 @@
 module Commands.Dev.Nockma.Run.Options
-  ( module Commands.Dev.Nockma.Run.BuiltinClient.Options,
+  ( module Commands.Dev.Nockma.Run.BuiltinEvaluator.Options,
     module Commands.Dev.Nockma.Run.EphemeralClient.Options,
     module Commands.Dev.Nockma.Run.WithClient.Options,
     module Commands.Dev.Nockma.Run.Options,
   )
 where
 
-import Commands.Dev.Nockma.Run.BuiltinClient.Options
+import Commands.Dev.Nockma.Run.BuiltinEvaluator.Options
 import Commands.Dev.Nockma.Run.EphemeralClient.Options
 import Commands.Dev.Nockma.Run.WithClient.Options
 import CommonOptions
 
 data NockmaRunCommand
-  = NockmaRunBuiltinClient NockmaRunBuiltinClientOptions
+  = NockmaRunBuiltinEvaluator NockmaRunBuiltinEvaluatorOptions
   | NockmaRunEphemeralClient NockmaRunEphemeralClientOptions
   | NockmaRunWithClient NockmaRunWithClientOptions
   deriving stock (Data)
@@ -23,17 +23,17 @@ parseNockmaRunCommand :: Parser NockmaRunCommand
 parseNockmaRunCommand =
   hsubparser
     ( mconcat
-        [ commandRunBuiltinClient,
+        [ commandRunBuiltinEvaluator,
           commandRunEphemeralClient,
           commandRunWithClient
         ]
     )
 
-commandRunBuiltinClient :: Mod CommandFields NockmaRunCommand
-commandRunBuiltinClient =
+commandRunBuiltinEvaluator :: Mod CommandFields NockmaRunCommand
+commandRunBuiltinEvaluator =
   command "builtin-evaluator" $
     info
-      (NockmaRunBuiltinClient <$> parseNockmaRunBuiltinClientOptions)
+      (NockmaRunBuiltinEvaluator <$> parseNockmaRunBuiltinEvaluatorOptions)
       (progDesc "Run with the builtin Nockma evaluator")
 
 commandRunEphemeralClient :: Mod CommandFields NockmaRunCommand
